requirements.txt
Flask
gunicorn
Flask-SSLify
connexion[swagger-ui]

psycopg2-binary
Flask-Script
Flask-Migrate
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
marshmallow
